+++ src/test/java/com/amazonaws/encryptionsdk/kmssdkv2/KMSProviderBuilderIntegrationTests.java
@@ -1,7 +1,7 @@
 // Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.amazonaws.encryptionsdk.kmsv2;
+package com.amazonaws.encryptionsdk.kmssdkv2;
 
 import static com.amazonaws.encryptionsdk.TestUtils.assertThrows;
 import static org.junit.Assert.*;
