+++ src/test/java/com/amazonaws/encryptionsdk/kmssdkv2/MockKmsClient.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.amazonaws.encryptionsdk.kmsv2;
+package com.amazonaws.encryptionsdk.kmssdkv2;
 
 import java.nio.ByteBuffer;
 import java.security.SecureRandom;
