+++ src/test/java/com/amazonaws/encryptionsdk/kmssdkv2/ProxyKmsClient.java
@@ -1,4 +1,4 @@
-package com.amazonaws.encryptionsdk.kmsv2;
+package com.amazonaws.encryptionsdk.kmssdkv2;
 
 import software.amazon.awssdk.awscore.exception.AwsServiceException;
 import software.amazon.awssdk.core.exception.SdkClientException;
